OpenAI Server refactoring (#2360)
