+++ requirements-dev.txt
@@ -16,3 +16,6 @@ pytest-asyncio
 httpx
 einops # required for MPT
 flash_attn # required for HuggingFace's llama implementation
+openai
+requests
+ray
